+++ web/app/features/auth/components/sign-up-card.tsx
@@ -6,7 +6,7 @@ import { Separator } from "../../../../components/ui/separator";
 import { TriangleAlert } from "lucide-react";
 import { v4 as uuidv4 } from "uuid";
 import usersData from "../../../../data/users.json";
-
+import React from "react";
 type Users = Record<string, { name: string; password: string; uid: string }>;
 const users = usersData as Users;
